Unify observation wrappers, use torch
